+++ .gitignore
@@ -3,3 +3,4 @@ venv/
 __pycache__/
 */__pycache__/
 dist/
+temp/
