+++ merkleasy/__init__.py
@@ -1,1 +1,2 @@
 from .classes import Tree
+from .errors import ImplementationError, SecurityError, UsagePreconditionError
