+++ tests/context.py
@@ -4,3 +4,4 @@ sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')
 
 import merkleasy
 from merkleasy import classes
+from merkleasy import errors
